Completely ignore invalid values
Before, the code assumed that a meaningful lux value could still be
deduced from invalid sensor values. This is not the case. It seems that
the invalid flag indicates an internal value overflow or saturation
in the sensor, which might come from a too high gain or a too long
integration time. In both cases one or both ADC values are wrong and
cannot be used.